Pre-compute spline once when bin contents are constant
Fixes #17.
